new draft for implementation of fib()
the basic idea:

fib() calls create an asymmetric tree, so we'll recursively call all the way from fib(n), fib(n-1), to fib(3). Critically, we'll pop all fib(n-2) values to the stack so they can be evaluated in the same manner at a later time. When fib() reaches fib(3) or fib(2), a counter register is incremented by 2 or 1, respectively (corresponding to the third and second fibonacci numbers), and the stack is checked for values to be tested.